requirements.txt
# requirements.txt
Django==2.0.3
pyyaml
dj-database-url
